Changed input fields & error logic
HTML fields now represent the proper requested data, acting as in-built verification for if an input is in fact a number. Errors will now appear as browser alerts, though this may change again during development. Currently, while HTML inputs are tagged with required and with autocomplete disabled, the testing browser (Firefox) still will allow incomplete forms to be submitted and will still autocomplete from the local cache. Not certain what's causing this yet, but it will be the focus of the next commit.