Added deductions/income
The application is now able to track deductions and income separately, rounding out the initial design goal of the project. The system simply checks the generated object's type property, which is determined by one of two radio buttons the user can click before submitting, and deducts or adds the amount to the user's total based on that value. Color coding was added to make entries easier to read, and may be added to the top total amount if deemed important enough for readability. Next update will be adding a delete button to existing entries in order to ease user organization.