small adjustments
Realized during development that the system already distinguishes between edited entries and originals with the differentation of the currentExp and entryObj objects. Because of this, I don't see too much benefit in creating another check to see if two entries have the same name or id, as a user may always delete duplicates if they wish to. Thus, this commit just brings with it some small tweaks, though a larger commit to rename variables is badly needed.