Made small adjustments & bugfix
Changed variable and html id's so that they make more sense with what the project currently has developed into. Additionally fixed a bug where deleting one element would delete the entire array. Had forgotten that javascript's splice method will delete all proceeding elements in an array if a second argument isn't given. So, I simply added a '1' as the second argument; kind of silly but that's how it is...